codes/python/certificates_data_entry_tools/pyproject.toml
[project]
name = "certificates-data-entry-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-connector-python>=9.5.0",
    "pandas>=3.0.0",
    "python-slugify>=8.0.4",
    "sqlalchemy>=2.0.46",
    "tabulate>=0.9.0",
]
